+++ src/lib/verify/findOtherType.ts
@@ -13,6 +13,7 @@ const supportedTypes: SupportedType[] = [
   'class',
   'interface',
   'program',
+  'transformation',
   'functionGroup',
   'serviceDefinition',
   'serviceBinding',